zset: pop the whole ZPOPMIN/ZPOPMAX window in one pass
The pop loop removed one element per iteration, and each removal cost a dict
lookup, a root-to-leaf descent and a possible rebalance, plus a copy of the
member that was needed only because the element was freed before its reply
was written.

The popped elements always form a rank window at one end of the tree, so the
whole reply can be emitted by walking the leaf chain and the removal done
with a single zbtDeleteRangeByRank() call, which unlinks a leaf slice at a
time. The reply layer copies members out, so nothing has to be duplicated to
survive the removal.

The keyspace event now fires once after the removal rather than after the
first element, which matches what the range-deletion commands already do and
leaves the observable event stream unchanged (verified against the listpack
encoding: one zpopmin, then del). The dict resize is paused for the batch and
shrunk once at the end, again as in the range-deletion path. The listpack
branch keeps its element-at-a-time loop.

zbtDeleteRangeByRank() takes unsigned long ranks now; the caller in
zremrangeGenericCommand() was already passing longs into unsigned int
parameters, which truncated above four billion elements.

+++ src/server.h
@@ -3832,7 +3832,7 @@ zbtElem *zbtFirstInLexRange(zbtree *t, zlexrangespec *range, unsigned long *out_
 zbtElem *zbtLastInLexRange(zbtree *t, zlexrangespec *range, unsigned long *out_rank, zbtIter *it);
 unsigned long zbtDeleteRangeByScore(zbtree *t, zrangespec *range, dict *d);
 unsigned long zbtDeleteRangeByLex(zbtree *t, zlexrangespec *range, dict *d);
-unsigned long zbtDeleteRangeByRank(zbtree *t, unsigned int start, unsigned int end, dict *d);
+unsigned long zbtDeleteRangeByRank(zbtree *t, unsigned long start, unsigned long end, dict *d);
 void zbtReplaceElem(zbtree *t, zbtElem *olde, zbtElem *newe);
 void zbtDefragNodes(zbtree *t, void *(*fn)(void *));
 int zbtDefragNodesIncremental(zbtree *t, void *(*fn)(void *), unsigned int budget);
